Set node 7.6 as latest version + split TokensData into own file
