+++ package.json
@@ -40,5 +40,8 @@
     "jsdoc": "^3.6.3",
     "mocha": "^7.0.0",
     "nyc": "^15.0.0"
+  },
+  "engines": {
+    "node": ">=7.6"
   }
 }
